Port Meowcoin consensus onto Bitcoin Core v30.2
- All 9 patches: consensus params, block header (CBlockVersion/AuxPoW/ProgPow),
  PoW algorithms (X16R/X16RV2/KAWPOW/MEOWPOW/Scrypt), chainparams, validation,
  P2P protocol, difficulty (LWMA/DGW), RPC (AuxPoW miner), asset layer foundation
- Build: zero errors, zero warnings (MSVC 19.44, x64)
- Smoke test: regtest genesis connects, node reaches Done loading
- Includes VALIDATION_REPORT.md with PoW-on-load analysis and test plan

+++ src/rpc/blockchain.cpp
@@ -159,8 +159,8 @@ UniValue blockheaderToJSON(const CBlockIndex& tip, const CBlockIndex& blockindex
     int confirmations = ComputeNextBlockAndDepth(tip, blockindex, pnext);
     result.pushKV("confirmations", confirmations);
     result.pushKV("height", blockindex.nHeight);
-    result.pushKV("version", blockindex.nVersion);
-    result.pushKV("versionHex", strprintf("%08x", blockindex.nVersion));
+    result.pushKV("version", static_cast<int32_t>(blockindex.nVersion));
+    result.pushKV("versionHex", strprintf("%08x", static_cast<int32_t>(blockindex.nVersion)));
     result.pushKV("merkleroot", blockindex.hashMerkleRoot.GetHex());
     result.pushKV("time", blockindex.nTime);
     result.pushKV("mediantime", blockindex.GetMedianTimePast());
